Simple module to perform queries to the ruc data
A simple module that performs queries to the ruc data.

It uses sqlite, in order to user it we need to read all the data
and stores in a sqlite database, for this we has two helper methods,
like:

	var helper = require('./db_helper');
	helper.hasData(function(hasData) {
		if (!hasData) helper.loadDataFromSet(useIt);
		else useIt();
	}

Where useIt is a function that perfom some queries, the helper
module also provide a simple function `findByRuc` that allows
the user to get data by the ruc (max 3 rows).